Started ViewPoint method
Started outlining implementation of the ViewPoint method. Added functionality to the Ray and Shower classes to meet the demands of ViewPoint. Defined an abstract IntensityFunctor class which contains the GetIntensity method. This functor is used by the Shower class to determine the number of photons emitted at a particular time.