Various cleanup and fixes
General polishing of code completed earlier. Also did a bit of work with coordinate transformations. The world and detector coordinate systems have been set up so they both share an x-axis and an origin. The z-axis of the world frame is perpendicular to the surface of the earth, while the z-axis of the detector frame is parallel to the axis of the detector.