Cherenkov direction distribution
Started work on calculating fluorescence profile. Waiting for more information on the fluorescence yield to continue there. Also created functions to generate random Cherenkov photon directions. The index of refraction is now estimated based on the local atmospheric density (delta is proportional to the density). This improves calculations of the Cherenkov threshold energy.