Corrected problem with determining coordinates of first interaction
If we use a negative value for x_0, we can't determine the starting height. A first interaction depth of zero corresponds to a height of infinity in the exponential atmosphere, and a negative first interaction depth doesn't make physical sense. Added a new constant, start_tracking, which specifies the depth at which we should start tracking the motion of the shower. The value for x_0 is still used in the GH profile.