Fixed time profile error calculation
Previously an error of bin_size/sqrt(n) was used. It has now been updated to use the standard formula for the standard deviation with Sheppard's correction. This was implemented in the new PhotonCount::TimeError method. It is unclear how this will be used to weight points in the fit graph by the number of points.